Unify TIF generation and update rendering
