+++ resources/Make TIFs/render_pngs.py
@@ -27,6 +27,15 @@ TERRAIN_PNG_DIR = BASE / "terrainPNGs"
 COLOR_SCALES_PATH = SCRIPT_DIR / "color_scales.json"
 CS = json.load(open(COLOR_SCALES_PATH))
 
+# Only publish a subset of timestamps to the web frontend
+TIMESTAMPS = [
+    "2025-05-24T09:00:00", "2025-05-24T12:00:00", "2025-05-24T15:00:00", "2025-05-24T18:00:00",
+    "2025-05-25T09:00:00", "2025-05-25T12:00:00", "2025-05-25T15:00:00", "2025-05-25T18:00:00",
+    "2025-05-26T09:00:00", "2025-05-26T12:00:00", "2025-05-26T15:00:00", "2025-05-26T18:00:00",
+    "2025-05-27T09:00:00", "2025-05-27T12:00:00", "2025-05-27T15:00:00", "2025-05-27T18:00:00",
+    "2025-05-28T09:00:00", "2025-05-28T12:00:00", "2025-05-28T15:00:00", "2025-05-28T18:00:00",
+]
+
 # Terrain TIF file mappings
 TERRAIN_TIFS = {
     "elevation": SCRIPT_DIR / "../terrains/tirol_100m_float.tif",
@@ -47,6 +56,23 @@ def build_cmap(palette_hex):
             # Assume hex format
             colors.append(color)
     return LinearSegmentedColormap.from_list("pal", colors, N=256)
+
+def render_sqh_png(depth_tif: pathlib.Path, qual_tif: pathlib.Path, out_png: pathlib.Path, spec: dict):
+    """Combine powder depth (opacity) and quality (colour) into a single PNG."""
+    cmap = build_cmap(spec["palette"])
+    vmin, vmax = spec.get("min", 0), spec.get("max", 1)
+
+    with rasterio.open(depth_tif) as dsrc, rasterio.open(qual_tif) as qsrc:
+        depth = dsrc.read(1).astype(float)
+        qual = qsrc.read(1).astype(float)
+        mask = (depth == dsrc.nodata) | np.isnan(depth)
+        depth_norm = np.clip(depth / 255.0, 0, 1)
+        qual_norm = np.clip((qual / 255.0 - vmin) / (vmax - vmin), 0, 1)
+
+    rgba = cmap(qual_norm)
+    rgba[..., 3] = depth_norm
+    rgba[mask, 3] = 0.0
+    plt.imsave(out_png, rgba)
 
 def needs_update(tif_path: pathlib.Path, png_path: pathlib.Path, var_name: str) -> bool:
     """Check if PNG needs to be updated based on TIF modification time and color scale changes."""
@@ -108,10 +134,20 @@ def terrain_needs_update(png_path: pathlib.Path, var_name: str) -> bool:
 
 def tif_to_png(tif_path: pathlib.Path):
     var = tif_path.stem  # Just use the stem directly
-    
+
     spec = CS.get(var)
     if not spec:        # skip un-known variable names
         print(f"  ⚠️  No colour spec for {var} (from file {tif_path.name})")
+        return
+
+    # Special handling for SQH which combines powder depth and quality
+    if var == "sqh":
+        depth_path = tif_path.parent / "powder_depth.tif"
+        quality_path = tif_path.parent / "powder_quality.tif"
+        if not depth_path.exists() or not quality_path.exists():
+            print(f"  ⚠️  Missing powder_depth/quality for {tif_path.parent.name}")
+            return
+        render_sqh_png(depth_path, quality_path, tif_path.parent / "sqh.png", spec)
         return
     
     # PNG name will be: temperature_2m.png
@@ -182,7 +218,9 @@ def main():
     # Process weather/composite layers (timestamp-based TIFs)
     print("🌤️  Processing weather and composite layers...")
     for ts_dir in BASE.iterdir():
-        if not ts_dir.is_dir() or ts_dir.name == "terrainPNGs": 
+        if not ts_dir.is_dir() or ts_dir.name == "terrainPNGs":
+            continue
+        if ts_dir.name not in TIMESTAMPS:
             continue
         for tif in ts_dir.glob("*.tif"):
             var = tif.stem  # Get variable name from filename
